+++ contracts/mocks/TestSTOFactory.sol
@@ -14,7 +14,10 @@ contract TestSTOFactory is IModuleFactory {
     constructor (address _polyAddress, uint256 _setupCost, uint256 _usageCost, uint256 _subscriptionCost) public
       IModuleFactory(_polyAddress, _setupCost, _usageCost, _subscriptionCost)
     {
-
+        version = "1.0.0";
+        name = "TestSTO";
+        title = "Test STO";
+        description = "Test STO";
     }
 
      /**
@@ -44,22 +47,30 @@ contract TestSTOFactory is IModuleFactory {
      * @notice Get the name of the Module
      */
     function getName() public view returns(bytes32) {
-        return "TestSTO";
+        return name;
     }
 
     /**
-     * @notice Get the description of the Module 
+     * @notice Get the description of the Module
      */
     function getDescription() public view returns(string) {
-        return "Test STO";
+        return description;
     }
 
     /**
      * @notice Get the title of the Module
      */
     function getTitle() public view returns(string) {
-        return "Test STO";
+        return title;
     }
+
+    /**
+     * @notice Get the version of the Module
+     */
+    function getVersion() public view returns(string) {
+        return version;
+    }
+
 
     /**
      * @notice Get the Instructions that helped to used the module