+++ contracts/modules/Checkpoint/ERC20DividendCheckpointFactory.sol
@@ -17,7 +17,8 @@ contract ERC20DividendCheckpointFactory is IModuleFactory {
      */
     constructor (address _polyAddress, uint256 _setupCost, uint256 _usageCost, uint256 _subscriptionCost) public
     IModuleFactory(_polyAddress, _setupCost, _usageCost, _subscriptionCost)
-    {
+    {   
+        version = "1.0.0";
         name = "ERC20DividendCheckpoint";
         title = "ERC20 Dividend Checkpoint";
         description = "Create ERC20 dividends for token holders at a specific checkpoint";
@@ -59,6 +60,13 @@ contract ERC20DividendCheckpointFactory is IModuleFactory {
      */
     function getTitle() public  view returns(string) {
         return title;
+    }
+
+    /**
+     * @notice Get the version of the Module
+     */
+    function getVersion() public view returns(string) {
+        return version;
     }
 
     /**