+++ contracts/modules/STO/DummySTOFactory.sol
@@ -15,7 +15,8 @@ contract DummySTOFactory is IModuleFactory {
      */
     constructor (address _polyAddress, uint256 _setupCost, uint256 _usageCost, uint256 _subscriptionCost) public
       IModuleFactory(_polyAddress, _setupCost, _usageCost, _subscriptionCost)
-    {
+    {   
+        version = "1.0.0";
         name = "DummySTO";
         title = "Dummy STO";
         description = "Dummy STO";
@@ -62,6 +63,13 @@ contract DummySTOFactory is IModuleFactory {
      */
     function getTitle() public view returns(string) {
         return title;
+    }
+
+    /**
+     * @notice Get the version of the Module
+     */
+    function getVersion() public view returns(string) {
+        return version;
     }
 
     /**