+++ contracts/modules/TransferManager/GeneralTransferManagerFactory.sol
@@ -14,7 +14,8 @@ contract GeneralTransferManagerFactory is IModuleFactory {
      */
     constructor (address _polyAddress, uint256 _setupCost, uint256 _usageCost, uint256 _subscriptionCost) public
       IModuleFactory(_polyAddress, _setupCost, _usageCost, _subscriptionCost)
-    {
+    {   
+        version = "1.0.0";
         name = "GeneralTransferManager";
         title = "General Transfer Manager";
         description = "Manage transfers using a time based whitelist";
@@ -53,6 +54,13 @@ contract GeneralTransferManagerFactory is IModuleFactory {
      */
     function getDescription() public view returns(string) {
         return description;
+    }
+
+    /**
+     * @notice Get the version of the Module
+     */
+    function getVersion() public view returns(string) {
+        return version;
     }
 
     /**