+++ contracts/modules/TransferManager/ManualApprovalTransferManagerFactory.sol
@@ -17,7 +17,8 @@ contract ManualApprovalTransferManagerFactory is IModuleFactory {
      */
     constructor (address _polyAddress, uint256 _setupCost, uint256 _usageCost, uint256 _subscriptionCost) public
       IModuleFactory(_polyAddress, _setupCost, _usageCost, _subscriptionCost)
-    {
+    {   
+        version = "1.0.0";
         name = "ManualApprovalTransferManager";
         title = "Manual Approval Transfer Manager";
         description = "Manage transfers using single approvals / blocking";
@@ -61,6 +62,13 @@ contract ManualApprovalTransferManagerFactory is IModuleFactory {
      */
     function getTitle() public view returns(string) {
         return title;
+    }
+    
+    /**
+     * @notice Get the version of the Module
+     */
+    function getVersion() public view returns(string) {
+        return version;
     }
 
     /**